Adding changes for MP5

+++ MP5/TopTitleStatisticsSpark.py
@@ -20,7 +20,7 @@ outputFile = open(sys.argv[2],"w")
 
 #TODO write your output here
 #write results to output file. Format
-ans1 = wc.sum()/wc.count()
+ans1 = int(wc.sum()/wc.count())
 outputFile.write('Mean\t%s\n' % ans1)
 ans2 = wc.sum()
 outputFile.write('Sum\t%s\n' % ans2)
@@ -28,7 +28,7 @@ ans3 = wc.min()
 outputFile.write('Min\t%s\n' % ans3)
 ans4 = wc.max()
 outputFile.write('Max\t%s\n' % ans4)
-ans5 = wc.variance()
+ans5 = int(wc.variance())
 outputFile.write('Var\t%s\n' % ans5)
